refactor: move generateRandomWord
