+++ packages/shared/lib/auxiliary/wordlists/index.ts
@@ -1,1 +1,2 @@
 export * from './english'
+export * from './utils'
