+++ packages/shared/lib/auxiliary/wordlists/utils/generateRandomWord.ts
@@ -1,0 +1,14 @@
+import { english } from '../english'
+
+export function generateRandomWord(excludedWords: string[]): string {
+    let word: string
+
+    do {
+        const wordIndex = Math.floor(Math.random() * english.length)
+        if (!excludedWords.includes(english[wordIndex])) {
+            word = english[wordIndex]
+        }
+    } while (!word)
+
+    return word
+}
