+++ packages/shared/lib/auxiliary/wordlists/utils/index.ts
@@ -1,0 +1,1 @@
+export * from './generateRandomWord'
